apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.19.0
python-multipart==0.0.9
pydantic==2.9.2
pillow==11.3.0
numpy>=1.24
# Optional for real OCR:
# paddleocr
# paddlepaddle
